requirements.txt
arcfutil==0.6.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.18.0
watchdog==2.1.3
Werkzeug==2.0.1
